Fixing GPU docker-compose for paddle
Adding Paddle Detect to the Worker docker image. Need the GPU version to work, but the coco-annotate code base is a few years old and the GPU version of the docker would not build anymore due to various lib version issues. I recreated the exact versions from the dockerhub stable versio and added the libs needed. With this version the webUI is working. Haven't tested the Paddle Detect tool yet.